% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staff.R
\name{staff}
\alias{staff}
\title{Staff details}
\usage{
staff(input, output = "output.html", ...)
}
\arguments{
\item{input}{a character string naming file which the tabular data are to be read from using
\code{\link[utils]{read.table}}. The following columns are mandatory: "name", "prefix", "degree",
"education", "interest", "email", "phone", "homepage", "lattes", and "image".}

\item{output}{a character string naming a HTML file for writing using \code{\link[base]{writeLines}}.}

\item{...}{arguments to \code{\link[base]{writeLines}}.}
}
\description{
Generate an HTML file containing data on the people working in our organization.
}
